+++ core/style.js
@@ -131,8 +131,15 @@ CKEDITOR.STYLE_OBJECT = 3;
 	 * @param {CKEDITOR.style} style
 	 */
 	CKEDITOR.editor.prototype.applyStyle = function( style ) {
-		if ( style.checkApplicable( this.elementPath() ) )
+		if ( style.checkApplicable( this.elementPath() ) ) {
+			var initialEnterMode = style._.enterMode;
+
+			// See comment in removeStyle.
+			if ( !initialEnterMode )
+				style._.enterMode = this.activeEnterMode;
 			applyStyleOnSelection.call( style, this.getSelection() );
+			style._.enterMode = initialEnterMode;
+		}
 	};
 
 	/**
@@ -142,8 +149,18 @@ CKEDITOR.STYLE_OBJECT = 3;
 	 * @param {CKEDITOR.style} style
 	 */
 	CKEDITOR.editor.prototype.removeStyle = function( style ) {
-		if ( style.checkApplicable( this.elementPath() ) )
+		if ( style.checkApplicable( this.elementPath() ) ) {
+			var initialEnterMode = style._.enterMode;
+
+			// There's no other way to pass editor's enter mode to the
+			// styles system and we need to do that (see #10190).
+			// However, we should not change style's enter mode if it was
+			// already set, because that could break backward compatibility.
+			if ( !initialEnterMode )
+				style._.enterMode = this.activeEnterMode;
 			applyStyleOnSelection.call( style, this.getSelection(), 1 );
+			style._.enterMode = initialEnterMode;
+		}
 	};
 
 	CKEDITOR.style.prototype = {